Replace tints with tones.
The v5 proposal included removing `oColorsGetTint` and renaming tints.
After some discussion with the Product Design team we decided not to
rename the palette tints black-50, teal-40, etc, as they're well known
and easy to remember. We decided to improve the documentation to explain
that black-[num], white-[num] is actually a mix not a tint.

However `oColorsGetTint` was restored for a new o-buttons usecase.
o-buttons also requires a function to take a colour name `claret-30` and
return the colour name and tint value `(color: claret, value: 30)`
so it can be incremented i.e. `oColorsGetTintDetails`. o-buttons expects
to use `oColorsGetTintDetails` for tints and another approach for mixes.
However `oColorsGetTint` can return a mix if a tint isn't supported.
If `oColorsGetTint` can return a colour which `oColorsGetTintDetails`
doesn't accept, that's confusing.

So:
- Make `oColorsGetTint` do one thing and error if it cannot return a tint
(no longer return mixes).
- But actually, a tint is a mixture of a color with white but our tints
are a function of saturation and lightness. So, since we are changing the
mixin behaviour also rename `oColorsGetTint` to `oColorsGetTone`.
- Add the new counterpart `oColorsGetToneDetails`.
- Rename tints to tones everywhere else, including demos.
- Keep existing palette colour names the same regardless of whether its a
tone (claret-50) or mix (black-50) -- again, as discussed with the
Product Design team, they're well known and easy to remember.

v5 proposal: https://github.com/Financial-Times/o-colors/issues/198
Tints and shades wiki: https://en.wikipedia.org/wiki/Tints_and_shades