Namespace colours and usecases with a slash in their name. (#226)
* Namespace colours and usecases with a slash in their name.

Setting a colour or colour usecase with colour name and project name
works well with o-colors, but is not great if you  want to pass a
custom colour name to a component like o-buttons. To define two colours
four arguments must be given (the colour name and where to get the colour
from).

Instead namespace colours and usecases using a forward slash
e.g. `o-social/twitter`.

```scss
// now
@include oColorsSetColor('o-social', 'twitter', blue);
@include oColorsSetUseCase('o-social', 'twitter-icon', (
   'background': ('color': 'twitter', 'from': 'o-social'),
   'text': 'white'
));
@include oButtonsAddTheme(
   $name: 'twitter',
   $opts: (
       'color': ('color': 'twitter', 'from': 'o-social'),
       'context': 'slate'
   ),
   $types: ('primary', 'secondary')
);

// udpated
@include oColorsSetColor('o-social/twitter', blue);
@include oColorsSetUseCase('o-social/twitter-icon', (
   'background': 'o-social/twitter',
   'text': 'white'
));
@include oButtonsAddTheme(
   $name: 'twitter',
   $opts: (
       'color': 'o-social/twitter',
       'context': 'slate'
   ),
   $types: ('primary', 'secondary')
);
```