expand the given hex colours to their full representation
The code in mixHexes assumes that the given hex colours will always be in their fully expanded form but this is not true.
This commit changes the mixHexes implementation to turn the given hex colours into their full form.

This fixes a bug where if mixHexes was given a hex colour such as fff then it would return NaN