Hide the B2C colour palette.
The B2C colour palette is not defined in the FTs Brand Guidelines
and is not defined in Customer Products' Figma foundations file.
It appears it is barely used.

We have found references in some components, e.g. o-buttons, o-header-services and the following
(non-achieved) repositories:
- https://github.com/Financial-Times/ip-lodestar-console/blob/master/client/styles.pcss#L38-L40
- https://github.com/Financial-Times/n-live-chat/blob/main/src/scss/_popup.scss#L19
- https://github.com/Financial-Times/next-tour-page/blob/main/client/community.scss#L132

See issue to deprecate B2C colours:
https://github.com/Financial-Times/o-colors/issues/316