Drop the `brand-` colour name prefix and update the readme.
FT Pink and FT Grey names were prefixed `brand-`, e.g. `brand-ft-pink`. I think we prefixed it because we were scared people would use it in digital products mistaking it for `paper`, and we wanted to make it distinct. I think that’s maybe unnecessary hand wringing (unnecessary worry) and not very helpful. We have documentation and design team members to make sure the right colour is used in the right place. Removing the prefix means improved code brevity (and autocompletion if we move to Sass variables) and means we can all use the same language from brand, through product design, to engineers.

`brand-ft-grey` is not deprecated as it is not yet released.